--- components/Home.js
@@ -1,13 +1,0 @@
-import React, { Component } from 'react';
-import Bus from './drawer/bus/bus';
-
-export default class Home extends Component {
-    constructor(props) {
-        super(props);
-
-    }
-
-    render = () => {
-        return <Bus />
-    }
-}
