--- components/drawer/settings/settingsDetail.js
@@ -1,12 +1,0 @@
-import React, { Component } from 'react'
-import { View, Text } from 'react-native'
-
-export default class extends Component {
-    render() {
-        return(
-            <View>
-                <Text>Detail</Text>
-            </View>
-        )
-    }
-}
