--- components/drawer/settings/settingsHome.js
@@ -1,14 +1,0 @@
-import React, { Component } from 'react'
-import { View, Text } from 'react-native'
-
-export default class extends Component {
-    render() {
-        return(
-            <View>
-                <Text>
-                    Home
-                </Text>
-            </View>
-        )
-    }
-}
