+++ components/navigations/navigationConfigs/index.js
@@ -1,0 +1,4 @@
+import foodTopTabNavigationConfig from './foodTopTabNavigationConfig';
+import stackNavigationConfig from './stackNavigationConfig';
+
+export { foodTopTabNavigationConfig, stackNavigationConfig }
