+++ components/navigations/stackNavigations/ScheduleStackNavigation.js
@@ -1,0 +1,12 @@
+import { stackNavigationOptions } from "../navigationConfigs";
+
+
+const ScheduleStackNavigator = createStackNavigator({
+    Schedule: {
+        screen: Schedule,
+    },
+    ScheduleDetail: {
+        screen: ScheduleDetail,
+    }
+}, stackNavigationOptions);
+
