+++ components/styles/busStyle.js
