+++ components/styles/colorPalette.js
@@ -1,0 +1,8 @@
+const colorPalette = {
+    mainColor: '#334955',
+    textColor: '#000000',
+    cardBackgroundColor: '#ffffff',
+    backgroundColor: '#f7f7f7',
+}
+
+export default colorPallette
